:recycle: Refactor: _cvtBGRA
